+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - Add guideline transparency control.
 - Add cancel action fallback to point mode.
+- Tweak curved fence mode to more closely match desired final point.
 - Fix continuous fence mode last item places for fences with asymmetrical Z-bounds (some UK pack props).
 
 ### 1.0.1.2